+++ Application/Persons/Queries/GetPersons/GetPersonsQueryValidator.cs
@@ -5,10 +5,10 @@ public class GetPersonsQueryValidator : AbstractValidator<GetPersonsQuery>
     public GetPersonsQueryValidator()
     {
         RuleFor(p => p.PageNumber)
-            .GreaterThanOrEqualTo(1);
+            .GreaterThanOrEqualTo((short)1);
 
         RuleFor(p => p.PageSize)
-            .GreaterThanOrEqualTo(10);
+            .GreaterThanOrEqualTo((short)10);
 
         RuleFor(p => p.Name)
             .MinimumLength(3)
