+++ Common/Wrappers/PaginatedResponse.cs
@@ -2,24 +2,26 @@
 
 public class PaginatedResponse<T>
 {
-    private PaginatedResponse(int pageIndex, int pageSize, int count, IEnumerable<T> items)
+    private PaginatedResponse(short pageIndex, short pageSize, short count, IEnumerable<T> items)
     {
         PageIndex = pageIndex;
-        TotalPages = (int)Math.Ceiling(count / (double)pageSize);
+        TotalPages = (short)Math.Ceiling(count / (double)pageSize);
         TotalCount = count;
         Items = items;
     }
 
-    public int PageIndex { get; }
-    public int TotalPages { get; }
-    public int TotalCount { get; }
+    
+    public short PageIndex { get; }
+    public short TotalPages { get; }
+    public short TotalCount { get; }
     public IEnumerable<T> Items { get; }
     public bool HasPreviousPage => PageIndex > 1;
     public bool HasNextPage => PageIndex < TotalPages;
     
-    public static async Task<PaginatedResponse<T>> CreateAsync(IQueryable<T> source, int pageIndex, int pageSize)
+    
+    public static async Task<PaginatedResponse<T>> CreateAsync(IQueryable<T> source, short pageIndex, short pageSize)
     {
-        var count = await source.CountAsync();
+        var count = (short)await source.CountAsync();
         var items = await source
             .Skip((pageIndex - 1) * pageSize)
             .Take(pageSize)
